Create clearly defined folder structure for backend.
