+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "wiki-crawl-to-mp3",
   "version": "1.0.0",
-  "main": "src/index.js",
+  "main": "src/server.js",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "node src/server.js",
     "test": "mocha --no-timeouts --exit \"./test/**/*.spec.js\""
   },
   "author": "Tom Edgeworth",
