+++ src/server.js
@@ -1,0 +1,8 @@
+require('dotenv').config();
+const app = require('./app');
+
+const port = process.env.PORT || 3000;
+
+app.listen(port, () => {
+  console.log(`Server is running on http://localhost:${port}`);
+});
