+++ src/utils/capitalizeWords.js
@@ -1,0 +1,6 @@
+module.exports = function capitalizeWords(input) {
+    return input
+      .split(' ')
+      .map(word => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
+      .join(' ');
+  };
