+++ src/utils/eventEmitter.js
@@ -1,0 +1,4 @@
+const events = require('events');
+const eventEmitter = new events.EventEmitter();
+
+module.exports = eventEmitter;
