+++ src/utils/streamToString.js
@@ -1,0 +1,7 @@
+module.exports = async function streamToString(stream) {
+    const chunks = [];
+    for await (const chunk of stream) {
+      chunks.push(chunk);
+    }
+    return Buffer.concat(chunks).toString('utf-8');
+  };
